OCR Image-to-Text Conversion/requirements.txt
pillow==11.0.0
pytesseract==0.3.13